+++ src/main/java/be/vdab/retrovideo/services/GenreService.java
@@ -6,6 +6,6 @@ import java.util.Optional;
 
 public interface GenreService {
     List<Genre> findAll();
-    Optional<Genre> findById(long id);
+    Optional<Genre> findById(int id);
     Optional<Genre> findByNaam(String naam);
 }
